+++ core/src/config.c
@@ -19,7 +19,7 @@ Config *create_config(const char *kwg_filename,
                       int game_pair_flag, int number_of_games_or_pairs,
                       const char *actual_tiles_played,
                       int player_to_infer_index, int actual_score,
-                      int number_of_tiles_exchanged, float equity_margin,
+                      int number_of_tiles_exchanged, double equity_margin,
                       int number_of_threads) {
 
   Config *config = malloc(sizeof(Config));
@@ -102,7 +102,7 @@ Config *create_config_from_args(int argc, char *argv[]) {
   int player_to_infer_index = -1;
   int actual_score = -1;
   int number_of_tiles_exchanged = 0;
-  float equity_margin = -1.0;
+  double equity_margin = -1.0;
   int number_of_threads = 1;
 
   int c;
